+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `setup.py` package list updated to include all subpackages in the wheel
 ## [0.5] - 2026-01-02
 ### Added
 - `DEEPSEEK_R1.LINEAR_ALL` model